refractor: rename files to index.* for consistency
