+++ apps/launch/components/MyStakingContracts/Details/index.tsx
