+++ apps/launch/pages/[network]/my-staking-contracts/[id].tsx
@@ -1,3 +1,3 @@
-import { Details } from 'components/MyStakingContracts/Details/Details';
+import { Details } from 'components/MyStakingContracts/Details';
 
 export default Details;
